Fix a bug with rspec-rails view specs not having helpers.
This line was causing my view specs to not have access to the default
set of Rails helpers. My current hypothesis is that prematurely
including cells/test_case is not the best idea, and causes some weird
loading issues.

This line should be moved into test_helper.rb or spec_helper.rb.

In the case of rspec-cells, Rails will automatically load
Cells::TestCase anyways, so we don't need this line.