Util::constant_for only searches descendants
Util::constant_for is used to by class_from_cell_name to construct a class name
from the cell name.

It uses const_get to resolve constants, which by default will resolve
ancestors before descendants.

This fix passes "inherit = false" to const_get to ensure that it only
searches descendants.

This restores the 4.0.5 behaviour of 'class_from_cell_name'.

Example:

The cell name is 'timesheets/client/overview', but Client is also the
name of an ActiveRecord model.

In 4.0.5 the resolved name would be Timesheets::Client::Overview
In 4.1.1 the resolved name is Clients::Overview (which is incorrect)
In 4.1.1 + this fix the resolved name is: Timesheets::Client::Overview